+++ examples/xcm-transfer/src/signers/v2/index.ts
@@ -1,3 +1,4 @@
 export * from './extension';
 export * from './provider';
-export * as asset from './types';
+export * from './utils';
+export * from './types';