+++ packages/xcm-cfg/src/chains/polkadot/polkadot.ts
@@ -5,7 +5,7 @@ import {
 
 import { dot } from '../../assets';
 
-export const polkadot = new Parachain({
+const config = {
   assetsData: [
     {
       asset: dot,
@@ -19,10 +19,22 @@ export const polkadot = new Parachain({
   explorer: 'https://polkadot.subscan.io',
   genesisHash:
     '0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3',
-  key: 'polkadot',
   name: 'Polkadot',
   parachainId: 0,
   ss58Format: 0,
-  trsry: '13UVJyLnbVp9RBZYFwFGyDvVd1y27Tt8tkntv6Q7JVPhFsTB',
+  treasury: '13UVJyLnbVp9RBZYFwFGyDvVd1y27Tt8tkntv6Q7JVPhFsTB',
+  usesDeliveryFee: true,
   ws: 'wss://polkadot-rpc.dwellir.com',
+};
+
+export const polkadot = new Parachain({
+  ...config,
+  key: 'polkadot',
+});
+
+export const polkadotCex = new Parachain({
+  ...config,
+  key: 'polkadot_cex',
+  name: 'Polkadot (CEX)',
+  usesCexForwarding: true,
 });