+++ packages/xcm-sdk/src/platforms/substrate/SubstratePlatform.ts
@@ -23,8 +23,9 @@ import {
 
 import { SubstrateService } from './SubstrateService';
 import { normalizeAssetAmount } from './utils';
+import { SubstrateCall } from './types';
 
-import { Platform, Call } from '../types';
+import { Platform } from '../types';
 
 export class SubstratePlatform
   implements Platform<ExtrinsicConfig, SubstrateQueryConfig>
@@ -37,18 +38,33 @@ export class SubstratePlatform
     this.#dex = DexFactory.getInstance().get(chain.key);
   }
 
+  private async useSignerFee(fee: Asset) {
+    const substrate = await this.#substrate;
+    return substrate.chain.usesSignerFee && !substrate.asset.isEqual(fee);
+  }
+
   async calldata(
     account: string,
     _amount: bigint,
+    feeBalance: AssetAmount,
     config: ExtrinsicConfig
-  ): Promise<Call> {
+  ): Promise<SubstrateCall> {
     const substrate = await this.#substrate;
+    const useSignerFee = await this.useSignerFee(feeBalance);
+
+    const txOptions = useSignerFee
+      ? {
+          asset: new Asset(feeBalance),
+        }
+      : undefined;
+
     const extrinsic = substrate.getExtrinsic(config);
     return {
       from: account,
       data: extrinsic.toHex(),
       type: CallType.Substrate,
-    } as Call;
+      txOptions: txOptions,
+    } as SubstrateCall;
   }
 
   async estimateFee(