+++ packages/xcm-sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @galacticcouncil/xcm-sdk
 
+## 8.1.1
+
+### Patch Changes
+
+- Fix mda account gen if dest relay
+- Updated dependencies
+  - @galacticcouncil/xcm-core@6.1.1
+
+## 8.1.0
+
+### Minor Changes
+
+- Substrate signer fee opts support
+- Swap re-work (fee swaps)
+- Automatic delivery fee calculation (based on treasury acc)
+
+### Patch Changes
+
+- Updated dependencies
+  - @galacticcouncil/xcm-core@6.1.0
+
 ## 8.0.1
 
 ### Patch Changes